+++ src/debug/light_wrapper.cpp
@@ -30,7 +30,7 @@ void register_light(game::Engine& engine, size_t id, const std::string& name)
 #endif
 }
 
-bool set_diffuse_color(game::Engine& engine, uint32_t id, const colorf& color)
+bool set_diffuse_color(game::Engine& engine, size_t id, const colorf& color)
 {
 	if (!engine.scene_context().light_pool.is_valid(id))
 		return false;
@@ -38,7 +38,7 @@ bool set_diffuse_color(game::Engine& engine, uint32_t id, const colorf& color)
 	return true;
 }
 
-bool get_diffuse_color(game::Engine& engine, uint32_t id, colorf& color)
+bool get_diffuse_color(game::Engine& engine, size_t id, colorf& color)
 {
 	if (!engine.scene_context().light_pool.is_valid(id))
 		return false;
@@ -46,7 +46,7 @@ bool get_diffuse_color(game::Engine& engine, uint32_t id, colorf& color)
 	return true;
 }
 
-bool set_specular_color(game::Engine& engine, uint32_t id, const colorf& color)
+bool set_specular_color(game::Engine& engine, size_t id, const colorf& color)
 {
 	if (!engine.scene_context().light_pool.is_valid(id))
 		return false;
@@ -54,7 +54,7 @@ bool set_specular_color(game::Engine& engine, uint32_t id, const colorf& color)
 	return true;
 }
 
-bool get_specular_color(game::Engine& engine, uint32_t id, colorf& color)
+bool get_specular_color(game::Engine& engine, size_t id, colorf& color)
 {
 	if (!engine.scene_context().light_pool.is_valid(id))
 		return false;
